requirements.txt
tensorflow==2.20.0
flask==2.3.3
flask-sqlalchemy==3.0.3
opencv-python==4.8.1.78
numpy==1.26.0
scikit-learn==1.3.0
matplotlib==3.7.2
pillow==10.0.1
flask-session
reportlab==4.1.0
scikit-image>=0.20.0
scipy==1.10.0
werkzeug==2.3.7
PyMySQL==1.1.0
cryptography==41.0.7
mysql-connector-python
python-dotenv==1.0.0
